Create a new package `org.wordpress.aztec.exceptions` and move DynamicLayoutGetBlockIndexOutOfBoundsException there
